src/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-multipart==0.0.9
SQLAlchemy==2.0.31
pyodbc==5.1.0
azure-identity==1.17.1
azure-storage-blob==12.23.1
azure-cosmos==4.7.0
azure-messaging-webpubsubservice==1.2.0
requests==2.32.3
httpx==0.27.0
orjson==3.10.6
python-multipart